docs :: 폴더 구조 변경 data, feature로 구분
