+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.DEMUANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".SplashActivity"
+            android:name=".feature.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,13 +22,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".OnBoardingActivity"
+            android:name=".feature.OnBoardingActivity"
             android:exported="false" />
-        <activity android:name=".MainActivity"
+        <activity android:name=".feature.MainActivity"
             android:exported="false" />
-        <activity android:name=".Login.LoginActivity"
+        <activity android:name=".feature.Login.LoginActivity"
             android:exported="false" />
-        <activity android:name=".signUp.SignUpActivity"
+        <activity android:name=".feature.signUp.SignUpActivity"
             android:exported="false" />
     </application>
 </manifest>
