--- app/src/main/java/com/example/demu_android/Blog/ModelBottomSheet.kt
@@ -1,23 +1,0 @@
-package com.example.demu_android.Blog
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.demu_android.R
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-class ModelBottomSheet : BottomSheetDialogFragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.list_bottom_sheet_option, container, false)
-    }
-
-    companion object {
-        const val TAG = "BasicBottomModalSheet"
-    }
-}
