+++ app/src/main/java/com/example/demu_android/data/api/ApiProvider.kt
@@ -1,0 +1,17 @@
+package com.example.demu_android.data.api
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object ApiProvider {
+    const val BASE_URL = ""
+
+    private val retrofit: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+
+}
