+++ app/src/main/java/com/example/demu_android/data/api/CommentApi.kt
@@ -1,0 +1,4 @@
+package com.example.demu_android.data.api
+
+interface CommentApi {
+}
