+++ app/src/main/java/com/example/demu_android/data/api/PostApi.kt
@@ -1,0 +1,4 @@
+package com.example.demu_android.data.api
+
+interface PostApi {
+}
