+++ app/src/main/java/com/example/demu_android/data/api/PublicAPi.kt
@@ -1,0 +1,4 @@
+package com.example.demu_android.data.api
+
+interface PublicAPi {
+}
