+++ app/src/main/java/com/example/demu_android/data/api/UserApi.kt
@@ -1,0 +1,5 @@
+package com.example.demu_android.data.api
+
+interface UserApi {
+    
+}
