+++ app/src/main/java/com/example/demu_android/feature/recycler/dum.kt
@@ -1,0 +1,4 @@
+package com.example.demu_android.feature.recycler
+
+class dum {
+}
