+++ app/src/main/java/com/example/demu_android/feature/recycler/home/WriteBlog/WriteBlogAdaptor.kt
@@ -1,0 +1,43 @@
+package com.example.demu_android.feature.recycler.home.WriteBlog
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.example.demu_android.databinding.ListBlogItemBinding
+import com.example.demu_android.feature.recycler.home.WriteBlog.data.Blog
+
+class
+WriteBlogAdaptor(
+    private val items: List<Blog>,
+) : RecyclerView.Adapter<WriteBlogAdaptor.Holder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
+        val binding = ListBlogItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+
+        return Holder(binding)
+    }
+
+    override fun getItemCount(): Int {
+        return items.size
+    }
+
+    override fun onBindViewHolder(holder: Holder, position: Int) {
+        holder.bind(items[position])
+    }
+
+    class Holder(
+        private val binding: ListBlogItemBinding
+    ) : RecyclerView.ViewHolder(binding.root) {
+//        init {
+//            binding.root.setOnClickListener {
+//                val position = adapterPosition
+//                if (position != RecyclerView.NO_POSITION) {
+//                    itemClickListener(position)
+//                }
+//            }
+//        }
+        fun bind(blog: Blog) {
+            binding.tvItemTitle.text = blog.title
+        }
+    }
+}
