+++ app/src/main/java/com/example/demu_android/feature/recycler/home/WriteBlog/data/Blog.kt
@@ -1,0 +1,9 @@
+package com.example.demu_android.feature.recycler.home.WriteBlog.data
+
+import com.example.demu_android.feature.type.Major
+
+data class Blog(
+    val title: String,
+    val content: String,
+    val major: Major
+)
