+++ app/src/main/java/com/example/demu_android/feature/recycler/home/dum.kt
@@ -1,0 +1,4 @@
+package com.example.demu_android.feature.recycler.home
+
+class dum {
+}
