+++ app/src/main/java/com/example/demu_android/feature/type/Major.kt
@@ -1,0 +1,5 @@
+package com.example.demu_android.feature.type
+
+enum class Major {
+    BACKEND, ANDROID
+}
