--- app/src/main/java/com/example/demu_android/recycler/dum.kt
@@ -1,4 +1,0 @@
-package com.example.demu_android.recycler
-
-class dum {
-}
