--- app/src/main/java/com/example/demu_android/recycler/home/WriteBlog/data/Blog.kt
@@ -1,10 +1,0 @@
-package com.example.demu_android.recycler.home.WriteBlog.data
-
-import com.example.demu_android.type.Major
-import com.example.demu_android.type.Tier
-
-data class Blog(
-    val title: String,
-    val content: String,
-    val major: Major
-)
