--- app/src/main/java/com/example/demu_android/recycler/home/dum.kt
@@ -1,4 +1,0 @@
-package com.example.demu_android.recycler.home
-
-class dum {
-}
