--- app/src/main/java/com/example/demu_android/type/Major.kt
@@ -1,5 +1,0 @@
-package com.example.demu_android.type
-
-enum class Major {
-    BACKEND, ANDROID
-}
