--- app/src/main/java/com/example/demu_android/type/Tier.kt
@@ -1,5 +1,0 @@
-package com.example.demu_android.type
-
-enum class Tier {
-    GRADE
-}
