+++ app/src/main/res/layout/activity_onboarding.xml
@@ -8,7 +8,7 @@
         android:background="@drawable/gradient"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".OnBoardingActivity">
+        tools:context=".feature.OnBoardingActivity">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/gl_top"
