+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,7 +7,7 @@
         android:background="@color/back_white"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Login.LoginActivity">
+        tools:context=".feature.Login.LoginActivity">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/glTop"
