+++ app/src/main/res/layout/activity_splash.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/back_white"
-    tools:context=".SplashActivity">
+    tools:context=".feature.SplashActivity">
 
     <ImageView
         android:layout_width="wrap_content"
