+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".home.HomeFragment">
+        tools:context=".feature.home.HomeFragment">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/gl_top"
