+++ app/src/main/res/layout/fragment_my_page.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Mypage.MyPageFragment">
+        tools:context=".feature.Mypage.MyPageFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
