+++ app/src/main/res/layout/fragment_write_blog.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/back_white"
-        tools:context=".Blog.WriteBlogFragment">
+        tools:context=".feature.Blog.WriteBlogFragment">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/gl_top"
